merge sections in Discussion
Merge sections "Our own limitations..." and "Why aren't more EEB...". I suggest that we merge these two sections as there is considerable overlap between the two and they compliment each other well. Probably worth checking with @vjf2 @Aguncan @shafieisabets that they are happy with this merging of the two sections in to one before we merge this PR however.