streamline Appendix S1 and S2 and code chunks (somewhat)
This commit edits S1 to be more concise, and moves some content from S2 into S1. This was done so the descriptions about how each index was computed is in one section. The Appendix S2 is now mainly the instructions to treat the data and plot it to generate Figure 2 from the main text. This code has also been edited down slightly, to only show the code needed to generate final Figure 2, without exercises in colour choices.